+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.10.5
+
+*Release Date*: 2020/1/8
+
+ ⚠️ **Deprecation Warning**: Direct state access in Operations is being removed. [Learn more](./DEPRECATION_NOTICE.md)
+
+- Add `StateProxy` to deprecate direct state access ([#128](https://github.com/Freshly/flow/pull/128))
+
 ## 0.10.4
 
 *Release Date*: 2020/1/7